// src/app/model/client.ts
export class Client {
       clientId: string;
    weight: number;
    leadClientId: number;
    dob: string;
    height: number;
    clientName: string;
    mobileNumber: string;
    emailId: string;
    endDate: string;
    active: number;
    age: number;
    bmi: number;
    aadharNumber: string;
    createdTime: string;
    isDeleted: number;
    activeUser: number;
    deletes: boolean;
    constructor() {
        
    }


}